+++ backend/canisters/openchat_installer/impl/src/updates/install_canisters.rs
@@ -6,6 +6,9 @@ use ic_cdk::management_canister::CanisterInstallMode;
 use ic_cdk::update;
 use openchat_installer_canister::CanisterType;
 use openchat_installer_canister::install_canisters::{Response::*, *};
+use p256_key_pair::P256KeyPair;
+use rand::rngs::StdRng;
+use rand::{Rng, SeedableRng};
 use types::{BuildVersion, C2CError, CanisterId, CanisterWasmBytes, Hash};
 use utils::canister::{
     CanisterToInstall, ChunkedWasmToInstall, WasmToInstall, clear_chunk_store, install, upload_wasm_in_chunks,
@@ -16,12 +19,15 @@ use utils::canister::{
 async fn install_canisters(args: Args) -> Response {
     let wasm_version = args.wasm_version;
 
+    let rng_seed = utils::canister::get_random_seed().await;
+    let mut rng = StdRng::from_seed(rng_seed);
+
     let PrepareResult {
         user_index,
         group_index,
         notifications_index,
         identity,
-    } = match read_state(|state| prepare(args, state)) {
+    } = match read_state(|state| prepare(args, &mut rng, state)) {
         Ok(ok) => ok,
         Err(response) => return response,
     };
@@ -54,7 +60,7 @@ struct InstallCanisterArgs<A: CandidType> {
     init_args: A,
 }
 
-fn prepare(args: Args, state: &State) -> Result<PrepareResult, Response> {
+fn prepare(args: Args, rng: &mut StdRng, state: &State) -> Result<PrepareResult, Response> {
     let user_index_wasm_hash = state.data.canister_wasms.chunks_hash(CanisterType::UserIndex);
     let group_index_wasm_hash = state.data.canister_wasms.chunks_hash(CanisterType::GroupIndex);
     let notifications_index_wasm_hash = state.data.canister_wasms.chunks_hash(CanisterType::NotificationsIndex);
@@ -69,7 +75,8 @@ fn prepare(args: Args, state: &State) -> Result<PrepareResult, Response> {
     } else if identity_wasm_hash != args.identity_wasm_hash {
         Err(HashMismatch(CanisterType::Identity, identity_wasm_hash))
     } else {
-        let ic_root_key = ic_cdk::api::root_key();
+        let oc_key_pair = P256KeyPair::new(rng);
+        let oc_secret_key_der = oc_key_pair.secret_key_der();
         let user_index = InstallCanisterArgs {
             canister_id: state.data.user_index_canister_id,
             wasm: state.data.canister_wasms.wasm_from_chunks(CanisterType::UserIndex),
@@ -92,7 +99,8 @@ fn prepare(args: Args, state: &State) -> Result<PrepareResult, Response> {
                 translations_canister_id: state.data.translations_canister_id,
                 website_canister_id: state.data.website_canister_id,
                 video_call_operators: args.video_call_operators.clone(),
-                ic_root_key: ic_root_key.clone(),
+                oc_secret_key_der: oc_secret_key_der.to_vec(),
+                rng_seed: rng.r#gen(),
                 wasm_version: args.wasm_version,
                 test_mode: state.data.test_mode,
             },
@@ -112,7 +120,7 @@ fn prepare(args: Args, state: &State) -> Result<PrepareResult, Response> {
                 registry_canister_id: state.data.registry_canister_id,
                 internet_identity_canister_id: state.data.internet_identity_canister_id,
                 video_call_operators: args.video_call_operators.clone(),
-                ic_root_key: ic_root_key.clone(),
+                rng_seed: rng.r#gen(),
                 wasm_version: args.wasm_version,
                 test_mode: state.data.test_mode,
             },
@@ -129,6 +137,7 @@ fn prepare(args: Args, state: &State) -> Result<PrepareResult, Response> {
                 authorizers: vec![state.data.user_index_canister_id, state.data.group_index_canister_id],
                 cycles_dispenser_canister_id: state.data.cycles_dispenser_canister_id,
                 registry_canister_id: state.data.registry_canister_id,
+                rng_seed: rng.r#gen(),
                 wasm_version: args.wasm_version,
                 test_mode: state.data.test_mode,
             },
@@ -145,7 +154,8 @@ fn prepare(args: Args, state: &State) -> Result<PrepareResult, Response> {
                 sign_in_with_email_canister_id: state.data.sign_in_with_email_canister_id,
                 originating_canisters: args.identity_originating_canisters,
                 skip_captcha_whitelist: args.identity_skip_captcha_whitelist,
-                ic_root_key: ic_root_key.clone(),
+                oc_secret_key_der: oc_secret_key_der.to_vec(),
+                rng_seed: rng.r#gen(),
                 wasm_version: args.wasm_version,
                 test_mode: state.data.test_mode,
             },