+++ backend/canisters/notifications_index/CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Use the endpoint as unique identifier for notification subscriptions ([#8670](https://github.com/open-chat-labs/open-chat/pull/8670))
 - Expose partial notification subscription details to aid with debugging ([#8671](https://github.com/open-chat-labs/open-chat/pull/8671))
 - Store date added alongside each notification subscription ([#8672](https://github.com/open-chat-labs/open-chat/pull/8672))
+- Take RNG seed in init args ([#8702](https://github.com/open-chat-labs/open-chat/pull/8702))
 
 ## [[2.0.1858](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1858-notifications_index)] - 2025-08-05